Event-based Puzzles
Making puzzles be owned by events, as a way to play with object
relationships.

The "CLICK HERE TO CHECK YOUR DATABASE SETUP" link now points to Events
instead of Puzzles, and the Events index has a new link to access that
event's puzzle list.

I've needed to pass the event ID around, and the event ID should
probably be part of some global context, but this is a helpful microcosm
of issues we'll have when doing other features so I still think it's a
good start.