Routing and event awareness cleanup (#57)
I've shrunk the work required for event awareness down to the very
small.

1. When declaring event-specific pages, derive page models from
ServerCore.ModelBases.EventSpecificModelBase. This adds an Event
property of type Event, and the menus will use it. Example:

public class DetailsModel : EventSpecificModelBase

2. Change the @page attribute to provide a specific routing path, e.g.
this one for puzzle details:

@page "{eventId}/Puzzles/Details/{id}"

[Note: we can probably remove this eventually, in favor of routing
conventions.]

3. When navigating from a page not in the event specific zone, add an
asp-route-eventId parameter, like this one in Events/Index that uses the
ID of the chosen event:

<a asp-page="/Puzzles/Index" asp-route-eventId="@item.ID">

That is currently the only place that asp-route-eventId is needed,
so these should be rare.