Prerequisite Editing
Allows prerequisites to be edited on the puzzle's Edit page. I've chosen to put them here rather than on their own page, which would have been a bit easier. This is the page they were edited on in the old site, and it was helpful to see the chosen prerequisites next to the min count to make sure that everything lined up.