Author versus Admin views on puzzles
Author and admin views are implemented for pages that show lists of puzzles - index, per-team status, and the map. The author view filters these lists, and also performs an access check before accessing the author/admin pages in the Puzzles folder.

Along the way, I standardized the constructor for EventSpecificPageModel - but I didn't add every single access check we'll need in the other folders.