EventId can be a friendly name instead of an int (#152)
* EventId can be a friendly name instead of an int

Now, if you supply a friendly name in the UrlString property, it will become the eventId. If you do not specify one, the ID will be used.

This isn't currently doing any kind of validation that the names are unique or anything. I also did not mark the new property as required because the migration was difficult. We can go back and make it required if we choose to.

* Review feedback