Edit Player Properties (#184)
Updated the profile page for the logged in user to update all their properties.

Also added a page that can be accessed from the team members list or the admin users list, which allows editing if the user is the current user or a global admin.

Along the way, fixed a bug where editing could clear the global admin flag.