Impersonation support
Using session cookies (not permanent ones) to allow global admins to impersonate any player. This could probably be enhanced to permit event-specific impersonation by event admins (by adding the event code to the cookie and validating that the impersonator is an event admin) but this will likely do for the short term.