Fix #416
Gather all the data we need in three queries and stitch the results together, rather than trust EF to build the proper query.

This logic is a rough analog to what's in the Map already, though the Map needs even more data than this so it's not an exact match.