General-purpose Intern Meta
This is just a straightforward piece viewer with no frills of any kind. It verifies that the event is an intern event, in case this facility could somehow be used to get an inappropriate look at an official hunt.

Event-specific instructions can be added by creating a piece whose progress level is <= 0.