Permit raw HTML in Responses
This lets authors put hyperlinks, emphasis, pictures, etc. in responses. There's a little bit of danger here but since this is only editable by authors and authors can only mangle their own puzzles, I think the expressivity is worth the danger. We to use this capability for the correct responses to the PD2020 metas.