Improve Import speed
The slowest part of Import turns out to be the file copy, by far. This change removes the file copy responsibilities from the import task entirely; the storage can simply be duplicated in Azure Storage Explorer by right-clicking and using the Clone... command, then providing the folder name that the import step adds to URLs for this event.