Improve Import Perf
Improve the perf of import such that it succeeds by omitting file copy, and provide user instructions for how to complete the job by clining the container in Azure Storage Explorer.

The model for the container display is very kludgy; I tried to do it the "right" way by creating a separate model item within Import.cshtml.cs, but SelectList was crashing internally and not explaining why. Apologies!