Add User Perf Fix
As the number of players has increased, the add-user list has gotten slower and slower. Making two attempts to fix it:

1. Requesting less data from the server (just ID/Name/Email, not all user fields)

2. Simplifying the query since we know that the set of users to skip will always be very small

If the true bottleneck is the amount of HTML transmitted to the client, this won't help, but that's unlikely since the page doesn't progressively render.

I also removed a Tuple<> from the add-team-member list that wasn't doing anything (Item2 was always -1).