Change All/Unsolved Filter from radio buttons to link (#749)
Radio buttons are traditionally used as selectors, not
as links.  Their behavior as links is unpredicatable,
and in this case only clicking on the text triggered
a navigation.  Clicking on the radio buttons did not,
which is confusing to the user.  Eliminate the confusion
by removing the radio button and showing a navigation link.

This affects the puzzles list and fastest solves pages. A
search of the codebase indicated that all other uses of
radio buttons were as input controls that did not trigger
navigation, so they were left as is.

Fixes #748