Team Announcements
Adds a TeamAnnouncement property on Event and a ShowTeamAnnouncement property on Team.

Announcement UX only shows when the event has an a TeamAnnouncement set.

When a team has its announcement turned on/off, an email is sent to all team members and an announcement displays on the top of every page.