Shared Resource UX
Allow an event to manage its shared resources in a folder on the Azure storage explicitly named "resources". This means that an event can drop any number of dependencies in that folder (CSS files, event logos, puzzle.js, etc) while keeping the URLs stable and not needing to use Azure Storage Explorer.