Unlocking Fixes
Adds a MinInGroupCount that can be used to unlock puzzles when enough puzzles in the Group have been solved.

Note that MinInGroupCount is nullable and the original MinPrerequisiteCount is not, so if you have a puzzle that you want to unlock based on group solves, you should set MinPrerequisiteCount to be some positive integer so it doesn't automatically unlock based on prerequisites. You can also set timed puzzles that count as prerequisites, and you should no longer need to make multiple copies of the same timer.

Also fixes the solve time of zero-prerequisite puzzles to be the event begin time instead of the current time.

Fixes #922.
Fixes #938.