New EphemeralHacks property for emergency usage
Whenever we have a property we want to temporarily add to an event, we can use this property on a short-term basis.

Intended usage is a space-delimited set of words that are ephemerally defined in our codebase. Anything with long-term importance should be moved out of this property and placed somewhere else.

In the very short term, we are going to use this to install temporary shutoff valves for some new features that may bog us down on the single front-end server that Puzzleday 2024 is going to have to run on.

Initial set of shutoff valves:
- kill-notifications: turns off toast notifications.
- kill-team-presence: turns off aggregated team presence on the puzzle list page.
- kill-presence: turns off all presence, both on the puzzle list page and on individual puzzles.