Fix returning closed notification issue (#1065)
The bug repros when multiple tabs are present, and each tab writes to localStorage based on its *current* state.

The correct thing to do is to load the notification list from localStorage and add the notification to that. If the player has many open tabs, those tabs will write the same value multiple times, and that's OK.

Fixes #1065.