Alpha Event Playtesting
Events can be marked as alpha testing events.

In an alpha testing event, authors can mark a puzzle with a count of how many alpha tests it needs.

As an alpha tester, the puzzle list page shows all puzzles that need alpha tests. When one of these puzzles is clicked, the puzzle content is grayed out unless/until the puzzle is claimed by the tester. Claiming a puzzle unlocks it for the tester and decrements the count.