Handle Login Drift
For years, we've been beset by a small number of users losing access to their accounts, needing manual reset at extremely inopportune times.

After a good deal of code review and web searching, I've finally discovered that a) this is normal and b) there is a simple bit of code that can be used for automatic and instantaneous account recovery.

I've tested this by mangling my local DB ProviderKey values, and observing that it successfully self-heals with the user being none the wiser.

Fixes #1073, though does it in a better way than expected.