Support role for puzzles
Allows the designation of individual users in a support role for a puzzle.

A puzzle supporter would function basically as an author who does not appear in the credits, so the safest course is to choose that path of least resistance and implement it that way. This minimizes the possibility for bugs.

Also taking this opportunity to broaden the animal representation of the demo project.