Html.Raw puzzle title support
Adding support for the raw syntax to a whole mess of scenarios. Should fix most/all issues in #1173 and a whole mess of unreported issues.

Note that a lot of these changes use puzzle.PlaintextName, which is already used in many places, defined as follows, and must be provided for this to work:

        /// The plaintext name of the puzzle, which is a subset of the name when the name is of the form:
        /// {plaintextName}Html.Raw({htmlName})